# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tashi API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  salesman: "salesman",
  mechanic: "mechanic",
  retailer: "retailer",
} as const;

export interface User {
  id: number;
  email: string;
  role: UserRole;
  points: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  salesman: "salesman",
  mechanic: "mechanic",
  retailer: "retailer",
} as const;

export interface CreateUserRequest {
  email: string;
  password: string;
  role: CreateUserRequestRole;
}

export interface Product {
  id: number;
  name: string;
  points: number;
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  points: number;
}

export type QRCodeStatus = (typeof QRCodeStatus)[keyof typeof QRCodeStatus];

export const QRCodeStatus = {
  unused: "unused",
  used: "used",
} as const;

export interface QRCode {
  id: number;
  qrNumber: string;
  productId: number;
  productName: string;
  points: number;
  status: QRCodeStatus;
  createdAt: string;
}

export interface CreateQRCodeRequest {
  qrNumber: string;
  productId: number;
}

export interface ScanQRCodeRequest {
  qrNumber: string;
}

export interface ScanResult {
  pointsEarned: number;
  totalPoints: number;
  productName: string;
  message: string;
}

export interface Scan {
  id: number;
  qrNumber: string;
  productName: string;
  pointsEarned: number;
  scannedAt: string;
}

export interface Ad {
  id: number;
  imageBase64: string;
  title?: string | null;
  createdAt: string;
}

export interface CreateAdRequest {
  imageBase64: string;
  title?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  salesmanId: number;
  retailerId: number;
  productId: number;
  quantity: number;
  totalPoints: number;
  bonusPoints: number;
  status: OrderStatus;
  productName?: string | null;
  retailerName?: string | null;
  retailerPhone?: string | null;
  createdAt: string;
}

export interface CreateOrderRequest {
  retailerId: number;
  productId: number;
  quantity: number;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface RetailerSummary {
  id: number;
  name?: string | null;
  phone: string;
  city?: string | null;
}

export interface BonusSummary {
  totalBonus: number;
  confirmedBonus: number;
  orders: Order[];
}

export type ListRetailersParams = {
  search?: string;
};