# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tashi API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current salesman (or all if admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (salesman only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/retailers:
    get:
      operationId: listRetailers
      tags: [orders]
      summary: Search retailers for order creation
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of retailers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RetailerSummary"
  /orders/my-bonus:
    get:
      operationId: getMyBonus
      tags: [orders]
      summary: Get salesman bonus summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bonus summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusSummary"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        points:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - points
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        points:
          type: integer
      required:
        - name
        - points
...
    Order:
      type: object
      properties:
        id:
          type: integer
        salesmanId:
          type: integer
        retailerId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        totalPoints:
          type: integer
        bonusPoints:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        productName:
          type: string
          nullable: true
        retailerName:
          type: string
          nullable: true
        retailerPhone:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - salesmanId
        - retailerId
        - productId
        - quantity
        - totalPoints
        - bonusPoints
        - status
        - createdAt
    CreateOrderRequest:
      type: object
      properties:
        retailerId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - retailerId
        - productId
        - quantity
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled]
      required:
        - status
    RetailerSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        phone:
          type: string
        city:
          type: string
          nullable: true
      required:
        - id
        - phone
    BonusSummary:
      type: object
      properties:
        totalBonus:
          type: integer
        confirmedBonus:
          type: integer
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required:
        - totalBonus
        - confirmedBonus
        - orders