# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tashi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderRequest {
  retailerId: number;
  productId: number;
  quantity: number;
}