# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tashi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  salesmanId: number;
  retailerId: number;
  productId: number;
  quantity: number;
  totalPoints: number;
  bonusPoints: number;
  status: OrderStatus;
  productName?: string | null;
  retailerName?: string | null;
  retailerPhone?: string | null;
  createdAt: string;
}